backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
numpy
pandas
fuzzywuzzy
websocket-client
schedule
matplotlib
